updating res project and build script

+++ build_RES.sh
@@ -1,15 +1,20 @@
 #!/bin/bash
 
+CHROME_BUILD=true
+FIREFOX_BUILD=false
+OPERA_BUILD=false
+
+
 ROOT=/Users/andytuba/Code/Reddit-Enhancement-Suite
-SRC=$ROOT/temp/ext
+SRC=$ROOT/dist #/temp/ext
 BUILD=$ROOT/temp/build
 PATH=$PATH:/usr/local/bin
 
 JETPACK=$ROOT/../addon-sdk
-FIREFOXSRC=$SRC/XPI
+FIREFOXSRC=$SRC/firefox
 FFEXTENSION=/Users/andytuba/Library/Application\ Support/Firefox/Profiles/06kw1qxd.default/extensions/jid1-xUfzOsOFlzSOXg@jetpack.xpi
 
-OPERASRC=$SRC/Opera
+OPERASRC=$SRC/opera  #todo: fix this yo
 OPERAEXTENSION=/Users/andytuba/Library/Application\ Support/Opera/widgets/reddit-enhancement-suite-2-4.1.5-1.oex
 
 SAFARISRC=$SRC/RES.safariextension
@@ -19,25 +24,25 @@ echo "Deploying RES to dev browsers from $SRC"
 
 cd $ROOT
 
-rm -rf $BUILD
-mkdir -p $BUILD
 
 #echo "Re-linking files from extension folders to lib/"
 #./makelinks.sh
-echo "Running grunt build"
-grunt
+echo "Gulping it down"
+gulp
 
 sleep .5
 
-
-open -a Google\ Chrome http://reload.extensions
-echo "Copied to Chrome"
-echo ""
-
+if $CHROME_BUILD; then
+	open -a Google\ Chrome http://reload.extensions
+	echo "Copied to Chrome"
+	echo ""
+fi
 
 #Clean
-rm $BUILD/opera.oex &> /dev/null
-if false; then
+
+if $OPERA_BUILD; then
+	#TODO: FIX THIS YO
+	rm $OPERABUILD/opera.oex &> /dev/null
 	#Build
 	cd $OPERASRC
 	zip -r $BUILD/opera.oex . -x '*.git/*' -x '*.DS_Store'
@@ -47,19 +52,22 @@ if false; then
 	echo ""
 fi
 
-# Clean
-rm -f $FIREFOXSRC/reddit_res.xpi &> /dev/null
-rm -f $BUILD/reddit_res.xpi &> /dev/null
-cd $JETPACK
-#Build
-source bin/activate
-cd $FIREFOXSRC
-cfx xpi #--force-mobile
-cp reddit_res.xpi $BUILD
-echo "Copied to Firefox"
-echo ""
+if $FIREFOX_BUILD; then
+	#todo: FIX THIS YO
 
+	# Clean
+	rm -f $FIREFOXSRC/reddit_res.xpi &> /dev/null
+	rm -f $BUILD/reddit_res.xpi &> /dev/null
+	cd $JETPACK
+	#Build
+	source bin/activate
+	cd $FIREFOXSRC
+	cfx xpi #--force-mobile
+	cp reddit_res.xpi $BUILD
+	echo "Copied to Firefox"
+	echo ""
 
-killall firefox &> /dev/null
-killall firefox-bin &> /dev/null
-cfx run
+	killall firefox &> /dev/null
+	killall firefox-bin &> /dev/null
+	cfx run
+fi
